pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.15.0"
fastapi = "^0.70.0"
psycopg2-binary = "^2.9.2"
alembic = "^1.7.5"
SQLAlchemy = "^1.4.27"
pyhumps = "^3.0.2"
gunicorn = "^20.1.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
sqlmodel = "^0.0.4"
fastapi-users = {extras = ["sqlalchemy"], version = "^8.1.2"}
sentry-sdk = "^1.5.0"
python-dotenv = "^0.19.2"
dramatiq = {extras = ["rabbitmq"], version = "^1.12.0"}
periodiq = "^0.12.1"
FastAPI-SQLAlchemy = "^0.2.1"


[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.11b1"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^6.2.5"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.11.29"
flake8-cognitive-complexity = "^0.1.0"
flake8-eradicate = "^1.2.0"
flake8-fixme = "^1.1.1"
flake8-isort = "^4.1.1"
flake8-mock = "^0.3"
flake8-multiline-containers = "^0.0.18"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^1.3.0"
flake8-print = "^4.0.0"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.5.1"
flake8-simplify = "^0.14.2"
flake8-todo = "^0.7"
flake8-use-fstring = "^1.3"
alembic-autogen-check = "^1.1.1"
gitlint = "^0.17.0"


[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

